:recycle: api: moving cpf-certification-result-repository.js to session bounded context
